Fin de l'implementation de la classe permanente employe et correction de certaine manquement dans les autres classe

+++ IPayment.cs
@@ -8,6 +8,6 @@ namespace YaoJosueAbotsidiaUA3Projet
 {
     public interface IPayment
     {
-        public abstract double CalculatePayment();
+        public abstract decimal CalculatePayment();
     }
 }
